finish proving beta reduction
